refactor(ton): move TonApi/TonApiDto to chains.ton sub-package
Groups all TON wire-format and client code under
`com.vultisig.wallet.data.api.chains.ton` so the parent `chains`
package isn't a dumping ground. Updates the 8 call sites
(repositories, status provider, use case, DI module, three test
files) to import from the new package.
