+++ data/src/main/kotlin/com/vultisig/wallet/data/repositories/BalanceRepository.kt
@@ -12,7 +12,7 @@ import com.vultisig.wallet.data.api.SolanaApi
 import com.vultisig.wallet.data.api.ThorChainApi
 import com.vultisig.wallet.data.api.TronApi
 import com.vultisig.wallet.data.api.chains.SuiApi
-import com.vultisig.wallet.data.api.chains.TonApi
+import com.vultisig.wallet.data.api.chains.ton.TonApi
 import com.vultisig.wallet.data.api.models.ResourceUsage
 import com.vultisig.wallet.data.api.models.calculateResourceStats
 import com.vultisig.wallet.data.api.models.thorchain.MergeAccount
